+++ src/test/java/com/lescastcodeurs/bot/ConferencesTest.java
@@ -1,0 +1,281 @@
+package com.lescastcodeurs.bot;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import java.util.List;
+import org.junit.jupiter.api.Test;
+
+class ConferencesTest {
+
+  @Test
+  void contentCanBeNull() {
+    Conferences confs = new Conferences(null, List.of());
+
+    assertTrue(confs.empty());
+    assertEquals("", confs.markdown());
+  }
+
+  @Test
+  void contentCanBeBlank() {
+    Conferences confs = new Conferences("   ", List.of());
+
+    assertTrue(confs.empty());
+    assertEquals("", confs.markdown());
+  }
+
+  @Test
+  void realContentWithoutCandidates() {
+    Conferences confs =
+        new Conferences(
+            """
+      # Developers Conferences Agenda/List
+
+      This repository lists a maximum of conference dates to help conference organizers.
+
+      As a conference organizer, we always ask ourselves what will be the ideal date for the event?
+
+      One piece of data to take into account is the date of other events, but how do you find it?
+
+      And as a developer, a technical lover, we want to have a calendar, we want to know when will be the next event.
+
+      I propose you this GitHub repository in order to list all the conferences for developers in France, Europe and the world.
+
+      You want to contribute, add an event or your event? Perfect! Just do a [Pull Request (PR)](https://github.com/scraly/developers-conferences-agenda/pulls) in order to update this developers conferences agenda.
+      If everybody contribute to this repo, it can really help conferences organizer and another goal of this idea is to list a maximum of conferences for developers.
+      If they don't know your conference they can't buy a ticket and come to it 😏.
+
+      This tech conferences agenda list can be seen in https://developers.events/
+
+      ## Archives
+
+      * [2017](archives/2017.md)
+      * [2018](archives/2018.md)
+      * [2019](archives/2019.md)
+      * [2020](archives/2020.md)
+      * [2021](archives/2021.md)
+      * [2022](archives/2022.md)
+
+      ## 2022
+
+      ### October
+
+      * 3-4: [KubeHuddle](https://www.kubehuddle.com/) - Edinburgh (Scotland) <a href="[https://conference-hall.io/public/event/wGVYkl21UFxeiuakhKfu](https://sessionize.com/kubehuddle/)"><img alt="Kube Huddle" src="https://img.shields.io/static/v1?label=CFP&message=until%2030-Jun-2022&color=red"> </a>
+      * 3-5: [NodeConf EU 2022](https://www.nodeconf.eu/) - Kilkenny (Ireland) <a href="[https://sessionize.com/nodeconf-eu-2022](https://sessionize.com/nodeconf-eu-2022)"><img alt="NodeConf EU 2022" src="https://img.shields.io/static/v1?label=CFP&message=until%2006-Jul-2022&color=red"> </a>
+      * 5-6: [PWA Summit 2022](https://www.pwasummit.org/) - Online <a href="https://sessionize.com/pwa-summit-2022"><img alt="PWA Summit 2022" src="https://img.shields.io/static/v1?label=CFP&message=until%2020-Jul-2022&color=red"> </a>
+      * 6-7: [Nordic.js](https://nordicjs.com/2022) - Stockholm (Sweden)
+      * 7-8: [Intersection Conference](https://www.intersection-conference.eu/) - Torino (Italy)
+
+      ### November
+
+      * 1-2: [ObservabilityCON](https://grafana.com/about/events/observabilitycon/2022) - New York City (US) & Online
+      * 2-3: [WordPress Accessibility Day](https://wpaccessibility.day/) - Online <img alt="Closed Captions" src="https://img.shields.io/static/v1?label=CC&message=Closed%20Captions&color=blue" /> <a href="https://wpaccessibility.day/participate/speaking-at-wp-accessibility-day/"><img alt="CFP WP Accessibility Day" src="https://img.shields.io/static/v1?label=CFP&message=until%2010-August-2022&color=red" /></a>
+      * 3: [JFall](https://jfall.nl/) - Netherlands
+      * 3: [Data Community Conference](https://www.data-community.ch) - Switzerland
+      * 3-4: [DevOps Barcelona](https://devops.barcelona/) - Barcelona (Spain) <a href="[https://sessionize.com/agile-tour-bordeaux-2022/](https://devops.barcelona/call-for-papers)"><img alt="CFP devops barcelona" src="https://img.shields.io/static/v1?label=CFP&message=until%2030-June-2022&color=red"> </a>
+      """,
+            List.of("(France)", "Devoxx"));
+
+    assertTrue(confs.empty());
+    assertEquals("", confs.markdown());
+  }
+
+  @Test
+  void realContentWithCandidates() {
+    Conferences confs =
+        new Conferences(
+            """
+      # Developers Conferences Agenda/List
+
+      This repository lists a maximum of conference dates to help conference organizers.
+
+      As a conference organizer, we always ask ourselves what will be the ideal date for the event?
+
+      One piece of data to take into account is the date of other events, but how do you find it?
+
+      And as a developer, a technical lover, we want to have a calendar, we want to know when will be the next event.
+
+      I propose you this GitHub repository in order to list all the conferences for developers in France, Europe and the world.
+
+      You want to contribute, add an event or your event? Perfect! Just do a [Pull Request (PR)](https://github.com/scraly/developers-conferences-agenda/pulls) in order to update this developers conferences agenda.
+      If everybody contribute to this repo, it can really help conferences organizer and another goal of this idea is to list a maximum of conferences for developers.
+      If they don't know your conference they can't buy a ticket and come to it 😏.
+
+      This tech conferences agenda list can be seen in https://developers.events/
+
+      ## Archives
+
+      * [2017](archives/2017.md)
+      * [2018](archives/2018.md)
+      * [2019](archives/2019.md)
+      * [2020](archives/2020.md)
+      * [2021](archives/2021.md)
+      * [2022](archives/2022.md)
+
+      ## 2022
+
+      ### October
+
+      * 3-4: [KubeHuddle](https://www.kubehuddle.com/) - Edinburgh (Scotland) <a href="[https://conference-hall.io/public/event/wGVYkl21UFxeiuakhKfu](https://sessionize.com/kubehuddle/)"><img alt="Kube Huddle" src="https://img.shields.io/static/v1?label=CFP&message=until%2030-Jun-2022&color=red"> </a>
+      * 3-5: [NodeConf EU 2022](https://www.nodeconf.eu/) - Kilkenny (Ireland) <a href="[https://sessionize.com/nodeconf-eu-2022](https://sessionize.com/nodeconf-eu-2022)"><img alt="NodeConf EU 2022" src="https://img.shields.io/static/v1?label=CFP&message=until%2006-Jul-2022&color=red"> </a>
+      * 4-6: [Devoxx Morocco](https://devoxx.ma/) - Agadir (Morocco) <a href="https://dvma22.cfp.dev"><img alt="CFP Devoxx Morocco" src="https://img.shields.io/static/v1?label=CFP&message=6-Jun-2022%20to%2015-Jul-2022&color=red"> </a>
+      * 5-6: [PWA Summit 2022](https://www.pwasummit.org/) - Online <a href="https://sessionize.com/pwa-summit-2022"><img alt="PWA Summit 2022" src="https://img.shields.io/static/v1?label=CFP&message=until%2020-Jul-2022&color=red"> </a>
+      * 6-7: [Nordic.js](https://nordicjs.com/2022) - Stockholm (Sweden)
+      * 6-7: [Paris Web](https://paris-web.fr/) - Paris (France) <img alt="Closed Captions" src="https://img.shields.io/static/v1?label=CC&message=Closed%20Captions&color=blue" /> <a href="https://appel.paris-web.fr/"><img alt="CFP Paris Web" src="https://img.shields.io/static/v1?label=CFP&message=24-Mar-2022-%3E24-Apr-2022&color=red"> </a>
+      * 7: [DevFest Perros-Guirec](https://devfest.codedarmor.fr/) - Perros-Guirec (France)
+      * 7: [Hackvens](https://hackvens.fr/) - Rennes (France)
+      * 7-8: [Intersection Conference](https://www.intersection-conference.eu/) - Torino (Italy)
+      * 10-14: [Devoxx Belgium](https://devoxx.be/) - Antwerp (Belgium) <a href="[https://devoxx.be/](https://dvbe22.cfp.dev/#/login)"><img alt="CFP Devoxx Belgium 2022" src="https://img.shields.io/static/v1?label=CFP&message=Jun-2022%20to%2015-July-2022&color=red"> </a>
+      * 12-13: [Kubernetes Community Days Munich 2022](https://community.cncf.io/events/details/cncf-kcd-munich-presents-kubernetes-community-days-munich-2022-1/) - Munich (Germany)
+      * 13-14: [Volcamp 2022](https://www.volcamp.io/) - Clermont Ferrand (France) <a href="https://conference-hall.io/public/event/8Wu147xoRDKQSjGcawoY"><img alt="CFP Volcamp" src="https://img.shields.io/static/v1?label=CFP&message=%3E13-May-2022&color=red"> </a>
+      * 13-14: [GeeCon Prague](https://www.geecon.org/) - Prague (Czechia)
+      * 13-14: [International Conference on Development and Learning](https://waset.org/development-and-learning-conference-in-october-2022-in-london) - London (United Kingdom)
+      * 13-14: [Forum PHP 2022](https://event.afup.org/) - Hotel New York - The Art of Marvel, Disneyland Paris (France) <img alt="Closed Captions" src="https://img.shields.io/static/v1?label=CC&message=Closed%20Captions&color=blue" /> <a href="https://afup.org/event/forumphp2022"><img alt="CFP Forum PHP 2022" src="https://img.shields.io/static/v1?label=CFP&message=until%2020-Jun-2022&color=red"> </a>
+      * 13-14: [EuroRust](https://eurorust.eu) - Berlin (Germany) & Online <a href="https://www.papercall.io/eurorust-2022"><img alt="CFP EuroRust" src="https://img.shields.io/static/v1?label=CFP&message=%3E28-August-2022&color=red"> </a>
+      * 14: [Fyne Conf](https://conf.fyne.io/) - Edinburgh (Scotland) & Online <a href="https://www.papercall.io/fyneconf2022"><img alt="CFP Fyne Conf" src="https://img.shields.io/static/v1?label=CFP&message=%3E09-Sept-2022&color=red"> </a>
+      * 16–20: [JavaOne](https://www.oracle.com/cloudworld/javaone/) - Las Vegas, Nevada (USA)
+      * 17: [Agile Tour Montpellier 2022](https://agiletourmontpellier.fr/2022/) - Montpellier (France)
+      * 18-19: [Dash](https://www.dashcon.io/) - New York (USA) <a href="https://www.dashcon.io/cfp/"><img alt="CFP Dash" src="https://img.shields.io/static/v1?label=CFP&message=until%2012-August-2022&color=red"> </a>
+      * 18-19: [FlowCon](https://flowcon.fr/) - Montrouge (France)
+      * 19-20: [BaselOne22](https://baselone.ch/one) - Basel (Switzerland) <a href="https://form.typeform.com/to/AYNtbnrl"><img alt="CFP BaselOne22" src="https://img.shields.io/static/v1?label=CFP&message=until%2018-Apr-2022&color=red"> </a>
+      * 19-20: [P99 Conf](https://www.p99conf.io/) - Virtual
+      * 20-21: [DevFest Nantes](https://devfest.gdgnantes.com/) - Nantes (France) <a href="https://conference-hall.io/public/event/CfucIjaXxrZThJE0POR3"> <img alt="Closed Captions" src="https://img.shields.io/static/v1?label=CC&message=Closed%20Captions&color=blue" /> <img alt="CFP Sunny Tech 2022" src="https://img.shields.io/static/v1?label=CFP&message=1-Apr-2022%20to%2019-Jun-2022&color=red"> </a>
+      * 20-21: [NixCon 2022](https://discourse.nixos.org/t/nixcon-2022-location-details-more/21392) - Paris (France)
+      * 24: [Open Observability Day (North America)](https://events.linuxfoundation.org/open-observability-day-north-america/) - Detroit (US)
+      * 24-27: [EclipseCon 2022](https://www.eclipsecon.org/2022) - Ludwigsburg (Germany) <a href="https://www.eclipsecon.org/2022/cfp"><img alt="CFP EclipseCon 2022" src="https://img.shields.io/static/v1?label=CFP&message=until%2015-Jun-2022&color=red"> </a>
+      * 24-28: [International JavaScript Conference Hybrid](https://javascript-conference.com/munich/) - Munich (Germany) & Online
+      * 24-28: [QCon San Francisco](https://qconsf.com/) - San Francisco (USA)
+      * 25-27: [SREcon22 Europe/Middle East/Africa](https://www.usenix.org/srecon) - Amsterdam (Netherlands)
+      * 26-27: [DevOpsDays Montreal](https://devopsdays.org/) - Montreal (Canada) <a href="https://docs.google.com/forms/d/e/1FAIpQLSf4qwC-lcdymPpvd15hUFqR6lVDABvXXcFNZrIwvNeugG7MgQ/viewform"><img alt="CFP DevOpsDays Montreal" src="https://img.shields.io/static/v1?label=CFP&message=until%2030-April-2022&color=red"> </a>
+      * 26-27: [API Days London](https://www.apidays.global/london/) - London (UK) & Online
+      * 26-27: [Arm DevSummit 2022](https://devsummit.arm.com/) - San Francisco (USA)
+      * 25-28: [KubeCon North America + CloudNativeCon](https://events.linuxfoundation.org/kubecon-cloudnativecon-north-america-2022/) - Detroit (USA)
+      * 27-28: [Agile Tour Bordeaux](http://agiletourbordeaux.fr/) - Bordeaux (France) <a href="https://sessionize.com/agile-tour-bordeaux-2022/"><img alt="CFP Agile Tour Bordeaux" src="https://img.shields.io/static/v1?label=CFP&message=until%2031-July-2022&color=red"> </a>
+      * 28: [Container Day](https://2022.containerday.it/) - Bologna (Italia)
+
+      ### November
+
+      * 1-2: [ObservabilityCON](https://grafana.com/about/events/observabilitycon/2022) - New York City (US) & Online
+      * 2-3: [WordPress Accessibility Day](https://wpaccessibility.day/) - Online <img alt="Closed Captions" src="https://img.shields.io/static/v1?label=CC&message=Closed%20Captions&color=blue" /> <a href="https://wpaccessibility.day/participate/speaking-at-wp-accessibility-day/"><img alt="CFP WP Accessibility Day" src="https://img.shields.io/static/v1?label=CFP&message=until%2010-August-2022&color=red" /></a>
+      * 3: [JFall](https://jfall.nl/) - Netherlands
+      * 3: [Data Community Conference](https://www.data-community.ch) - Switzerland
+      * 3-4: [DevOps Barcelona](https://devops.barcelona/) - Barcelona (Spain) <a href="[https://sessionize.com/agile-tour-bordeaux-2022/](https://devops.barcelona/call-for-papers)"><img alt="CFP devops barcelona" src="https://img.shields.io/static/v1?label=CFP&message=until%2030-June-2022&color=red"> </a>
+      * 3-4: [Agile Tour Nantes 2022](https://agilenantes.org/evenements/agile-tour-nantes/agile-tour-nantais-2022/) - Nantes (France)
+      * 8-9: [Open Source Experience](https://www.opensource-experience.com/) - Paris (France) <a href="https://sessionize.com/open-source-experience-2022/"><img alt="CFP Open Source Experience" src="https://img.shields.io/static/v1?label=CFP&message=until%2015-Jun-2022&color=red"> </a>
+      * 14-16: [OSMC](https://osmc.de/) - Nuremberg (Germany) <a href="https://osmc.de/call-for-papers/"><img alt="CFP OSMC" src="https://img.shields.io/static/v1?label=CFP&message=until%2031-Jul-2022&color=red"> </a>
+      * 15-16: [ParisTestConf](https://paristestconf.com/) - Online <a href="https://paristestconf.com/call-for-papers-2022"><img alt="CFP ParisTestConf" src="https://img.shields.io/static/v1?label=CFP&message=until%2031-May-2022&color=red"> </a>
+      * 15-16: [Agile Tour Toulouse](https://tour.agiletoulouse.fr/) - Toulouse (France) <a href="https://sessionize.com/agile-tour-toulouse-2022/"><img alt="CFP Agile Tour Toulouse" src="https://img.shields.io/static/v1?label=CFP&message=until%2026-Jun-2022&color=red"> </a>
+      * 16-17: [NODES 2022 (Neo4j Online Developer Expo and Summit)](https://neo4j.com/nodes-2022/) - Online <a href="https://sessionize.com/neo4j-nodes-2022/"><img alt="Neo4j Online Developer Expo and Summit)" src="https://img.shields.io/static/v1?label=CFP&message=until%2020-Aug-2022&color=red"> </a>
+      * 17: [Codeurs en Seine](https://www.codeursenseine.com/) - Rouen (France) <a href="[https://sessionize.com/agile-tour-toulouse-2022/](https://conference-hall.io/public/event/Uv6XULhgmt7z2YuoJ8nq)"><img alt="CFP codeur en seine" src="https://img.shields.io/static/v1?label=CFP&message=until%2016-Sept-2022&color=red"> </a>
+      * 17: [Kubernetes Community Days Italy 2022](https://community.cncf.io/events/details/cncf-kcd-italy-presents-kubernetes-community-days-italy-2022/) - Milan (Italy)
+      * 17: [lbc² by leboncoin](https://www.lbc2.fr/) - Paris (France)
+      * 18: [DevFest Strasbourg](https://devfest.gdgstrasbourg.fr/) - Strasbourg (France) <a href="https://conference-hall.io/speaker/event/wgaja45IBdGT4sw4CXDB"><img alt="CFP Devfest Strasbourg" src="https://img.shields.io/static/v1?label=CFP&message=until%2030-Jun-2022&color=red"> </a>
+      * 18-19: [DevOpsDays Bengaluru](https://devopsdaysindia.org/) - Bengaluru (India) <a href="https://forms.gle/LmCGgZse3Lp4XNgz9"><img alt="CFP DevOpsDays Bengaluru" src="https://img.shields.io/static/v1?label=CFP&message=until%206-Aug-2022&color=red"> </a>
+      * 18-19: [GreHack](https://grehack.fr/) - Grenoble (France) <a href="https://easychair.org/cfp/GreHack2022"><img alt="CFP" src="https://img.shields.io/static/v1?label=CFP&message=until%2015-Aug-2022&color=red"> </a>
+      * 19-20: [Capitole du Libre](https://capitoledulibre.org/) - Toulouse (France) <a href="https://cfp.capitoledulibre.org/cdl-2022/cfp"><img alt="CFP Capitole du Libre de Toulouse" src="https://img.shields.io/static/v1?label=CFP&message=until%2026-Sep-2022&color=red"> </a>
+      * 21: [Java Summit-IL](https://www.javasummitil.com/) - Tel-Aviv (Israel) <a href="https://sessionize.com/javasummit-il-22/"><img alt="CFP Java Summit-IL" src="https://img.shields.io/static/v1?label=CFP&message=until%2031-Jul-2022&color=red"> </a>
+      * 22-23: [Kubernetes Community Days UK 2022](https://community.cncf.io/events/details/cncf-kcd-uk-presents-kubernetes-community-days-uk-2022/) - London (UK)
+      * 22-24: [Kubernetes Community Days Spain 2022](https://community.cncf.io/events/details/cncf-kcd-spain-presents-kcd-spain-2022/) - Spain
+      * 23-25: [Agile Grenoble 2022](http://agile-grenoble.org/) - Grenoble (France)
+      * 26-27: [Kubernetes Community Days & OpenInfra Days Indonesia 2022](https://community.cncf.io/events/details/cncf-kcd-indonesia-presents-kubernetes-community-days-openinfra-days-indonesia-2022/) - Jakarta (Indonesia)
+      * 29-9/12: [QCon Plus](https://plus.qconferences.com/) - Online
+
+      ### December
+
+      * 1: [Devops DDay #7](https://2022.devops-dday.com/) - Marseille (France) <a href="https://conference-hall.io/public/event/eKYGzptI6y44zoS8sGUz"><img alt="CFP Devops DDay" src="https://img.shields.io/static/v1?label=CFP&message=until%2015-July-2022&color=red"> </a>
+      * 2: [BDX I/O](https://www.bdxio.fr/) - Bordeaux (France) <a href="https://conference-hall.io/public/event/OJC5Ou5YJodfetgSJCa3"><img alt="CFP BDX I/O" src="https://img.shields.io/static/v1?label=CFP&message=until%2031-Jul-2022&color=red"> </a>
+      * 2: [DevFest Dijon 2022](https://www.linkedin.com/posts/developers-group-dijon_apr%C3%A8s-des-%C3%A9changes-avec-google-le-developers-activity-6973555910140903424-rjRH/) - Dijon (France) <a href="[https://conference-hall.io/public/event/ZGJWM1x64evaNUzycEzn](https://conference-hall.io/public/event/cXWCepNzXVWwb1i6ktVA)"><img alt="CFP DevFest Dijon 2022" src="https://img.shields.io/static/v1?label=CFP&message=until%2014-October-2022&color=green"> </a>
+      * 2-3: [Oπe\\n Conf Athens 2022](https://www.open-conf.gr/) - Athens (Greece) <a href="https://sessionize.com/open_conf/"><img alt="CFP Oπe\\n Conf Athens 2022" src="https://img.shields.io/static/v1?label=CFP&message=until%2031-Aug-2022&color=red"> </a>
+      * 5-8: [DevOpsCon Munich](https://devopscon.io/call-for-papers/) - Munich (Germany) <a href="https://devopscon.io/call-for-papers/"><img alt="CFP DevOpsCon Munich" src="https://img.shields.io/static/v1?label=CFP&message=until%2027-Jun-2022&color=red"> </a>
+      * 6-7: [DevRel Con Prague](https://github.com/scraly/developers-conferences-agenda/) - Prague (Czech Republic)
+      * 7-9: [SREcon22 Asia/Pacific](https://www.usenix.org/srecon) - Sydney (Australia)
+      * 8: [Agile Tour Lausanne 2022](https://www.agilesuisse.ch/agile-tour-lausanne-2022) - Lausanne (Suisse)
+      * 14-15: [DevOpsDays Tel-Aviv](https://sessionize.com/devopsdays-tel-aviv-2022/) - Tel-Aviv <a href="https://sessionize.com/devopsdays-tel-aviv-2022/"><img alt="DevOpsDays Tel-Aviv" src="https://img.shields.io/static/v1?label=CFP&message=until%2031-August-2022&color=red"> </a>
+      * 14-16: [API Days Paris](https://www.apidays.global/paris/) - Paris (France) & Online
+      * 15-16: [Agile Tour Rennes](https://agiletour.agilerennes.org/) - Rennes (France) <a href="https://sessionize.com/agile-tour-rennes-2022/"><img alt="Agile Tour Rennes" src="https://img.shields.io/static/v1?label=CFP&message=until%2015-October-2022&color=green"> </a>
+
+      ## 2023
+
+      ### January
+
+      * 19-20: [Touraine Tech](https://touraine.tech/) - Tours (France) <a href="https://conference-hall.io/speaker/event/3lWSdH0pfZkHEAL7RWSJ"><img alt="Touraine Tech 2023" src="https://img.shields.io/static/v1?label=CFP&message=from%20September-30%20to%20November-27-2022&color=green"> </a>
+      * 23-27: [NDC London](https://ndclondon.com/) - London (UK) <a href="https://sessionize.com/ndc-london-2023"><img alt="CFP NDC London" src="https://img.shields.io/static/v1?label=CFP&message=until%2018-September-2022&color=red"> </a>
+      * 25-28: [SnowCamp](https://snowcamp.io/fr/) - Grenoble (France) <a href="https://conference-hall.io/public/event/ZGJWM1x64evaNUzycEzn"><img alt="CFP SnowCamp 2023" src="https://img.shields.io/static/v1?label=CFP&message=until%2015-October-2022&color=green"> </a>
+
+      ### February
+
+      * 4-5: [FOSDEM](https://fosdem.org/2022/news/2022-09-14-fosdem-2023-dates/) - Belgium
+      * 9-10: [Latam 2023 DevRelCon](https://latam-2023.devrelcon.dev/) - Online
+      * 22-24: [ConFoo](https://confoo.ca/fr/2023) - Montréal (Canada) <a href="https://confoo.ca/fr/2023/call-for-papers"><img alt="CFP ConFoo 2023" src="https://img.shields.io/static/v1?label=CFP&message=until%2023-September-2022&color=red"> </a>
+      * 23-24: [Kubernetes Community Days Amsterdam 2023](https://community.cncf.io/events/details/cncf-kcd-netherlands-presents-kubernetes-community-days-amsterdam-2023/) - Amsterdam (NL)
+
+      ### March
+
+      * 7: [Kubernetes Community Days France](https://community.cncf.io/events/details/cncf-kcd-france-presents-kubernetes-community-days-france-2023/) - Paris (France) <img alt="CFP KCD Paris" src="https://img.shields.io/static/v1?label=CFP&message=from%20September-15%20to%20October-31-2022&color=red">
+      * 16: [Voxxed Days Zurich](https://voxxeddays.com/zurich/) - Zurich (Switzerland)<img alt="CFP VDZ" src="https://img.shields.io/static/v1?label=CFP&message=from%20October-10%20to%20November-16-2022&color=green">
+      * 23: [DevDay](https://www.devday.be/) - Louvain la Neuve (Belgium)
+      * 27-29: [QCon London](https://qconlondon.com/) - London (UK)
+
+      ### April
+
+      * 12-14: [Devoxx France](https://www.devoxx.fr/) - Paris (France)
+      * 12-14: [KotlinConf](https://kotlinconf.com/) - Amsterdam (the Netherlands) <a href="https://sessionize.com/kotlinconf-2023/"><img alt="CFP KotlinConf" src="https://img.shields.io/static/v1?label=CFP&message=until%201-October-2022&color=red"> </a>
+      * 17-21: [KubeCon Europe + CloudNativeCon](https://events.linuxfoundation.org/kubecon-cloudnativecon-europe-2023/) - Amsterdam (THE NETHERLANDS)
+      * 24-25: [DevOps Days Geneva](https://devopsdays.org/events/2022-geneva/welcome/) - Geneva (Switzerland)
+
+      ### May
+      * ??: [Codecamp_Festival](https://codecamp.ro/the-festival-book-now/) - Bucharest, Romania
+      * 23: [VoxxedDays Brussels](https://brussels.voxxeddays.com/) - Brussels (Belgium)
+
+      ### June
+
+      * 20-23: [JCON PEOPLE](https://jcon.one/) - Düsseldorf, Germany\s
+
+      ### July
+
+      * 3-5: [DevBCN](www.devbcn.com) - Barcelona (Spain)
+
+      ### September
+
+      * 21-23: [Strange Loop 2023](https://thestrangeloop.com) - St. Louis, Missouri (USA)
+
+      ### October
+
+      * 2-6: [QCon San Francisco](https://qconsf.com/) - San Francisco (USA)
+      * 10-12: [SREcon23 Europe/Middle East/Africa](https://www.usenix.org/srecon) - Dublin (Ireland)
+      """,
+            List.of("(France)", "Devoxx"));
+    String markdown =
+        confs
+            .markdown()
+            // s.matches(regex) ne fonctionne pas sur des multiline strings
+            .replaceAll("\\n", " ");
+
+    assertFalse(confs.empty());
+    assertTrue(markdown.matches(".*4-6 octobre 2022 : .*Devoxx Morocco.*"));
+    assertTrue(markdown.matches(".*6-7 octobre 2022 : .*Paris Web.*"));
+    assertTrue(markdown.matches(".*15-16 novembre 2022 : .*Agile Tour Toulouse.*"));
+    assertTrue(markdown.matches(".*19-20 janvier 2023 : .*Touraine Tech.*"));
+  }
+
+  @Test
+  void contentWithInvalidMonth() {
+    Conferences confs =
+        new Conferences(
+            """
+      ## 2022
+
+      ### InvalidMonth
+
+      * 6-7: [Paris Web](https://paris-web.fr/) - Paris (France) <img alt="Closed Captions" src="https://img.shields.io/static/v1?label=CC&message=Closed%20Captions&color=blue" /> <a href="https://appel.paris-web.fr/"><img alt="CFP Paris Web" src="https://img.shields.io/static/v1?label=CFP&message=24-Mar-2022-%3E24-Apr-2022&color=red"> </a>
+      """,
+            List.of("(France)"));
+    String markdown =
+        confs
+            .markdown()
+            // s.matches(regex) ne fonctionne pas sur des multiline strings
+            .replaceAll("\\n", " ");
+
+    assertFalse(confs.empty());
+    assertTrue(markdown.matches(".*6-7 InvalidMonth 2022 : .*Paris Web.*"));
+  }
+}